+++ Modules/Intent.Modules.Common.Java/Builder/JavaField.cs
@@ -7,6 +7,7 @@ public class JavaField : JavaMember<JavaField>
     public string Type { get; }
     public string Name { get; }
     public string AccessModifier { get; private set; }
+    public string DefaultValue { get; private set; }
 
     public JavaField(string type, string name)
     {
@@ -25,9 +26,9 @@ public class JavaField : JavaMember<JavaField>
         Name = name;
     }
 
-    public JavaField Private()
+    public JavaField Public()
     {
-        AccessModifier = "private ";
+        AccessModifier = "public ";
         return this;
     }
 
@@ -42,9 +43,27 @@ public class JavaField : JavaMember<JavaField>
         AccessModifier = "private final ";
         return this;
     }
+    
+    public JavaField PublicFinal()
+    {
+        AccessModifier = "public final ";
+        return this;
+    }
+
+    public JavaField WithoutAccessModifier()
+    {
+        AccessModifier = "";
+        return this;
+    }
+    
+    public JavaField WithDefaultValue(string defaultValue)
+    {
+        DefaultValue = defaultValue;
+        return this;
+    }
 
     public override string GetText(string indentation)
     {
-        return $@"{GetComments(indentation)}{GetAnnotations(indentation)}{indentation}{AccessModifier}{Type} {Name};";
+        return $@"{GetComments(indentation)}{GetAnnotations(indentation)}{indentation}{AccessModifier}{Type} {Name}{(DefaultValue != null ? $" = {DefaultValue}" : string.Empty)};";
     }
 }