+++ Modules/Intent.Modules.Common.Java/release-notes.md
@@ -3,6 +3,7 @@
 - Updated Java File Builder in order for `final` to be applied in fields and parameters.
 - Made minor corrections in the naming of certain Java concepts in the File Builder.
 - Added Generic Type parameter support for Classes and Methods (no constraint support yet).
+- Modified the Java Fields and Interface Fields with appropriate defaults and with additional access modifier methods. Fields can also have default values.
 
 ### Version 3.4.3